+++ src/commonMain/kotlin/io/smallibs/aktor/Behavior.kt
@@ -2,7 +2,7 @@
 
 package io.smallibs.aktor
 
-import io.smallibs.aktor.core.Behaviors
+import io.smallibs.aktor.core.Core.Behaviors
 
 typealias Receiver<T> = (Actor<T>, Envelop<T>) -> Unit
 typealias CoreReceiver<T> = (Actor<T>, CoreEnvelop<T>) -> Unit
