--- src/commonMain/kotlin/io/smallibs/aktor/core/CoreMessage.kt
@@ -1,13 +1,0 @@
-package io.smallibs.aktor.core
-
-import io.smallibs.aktor.ActorReference
-
-interface Core {
-
-    interface Protocol
-    object Start : Protocol
-    object Stop : Protocol
-    data class Stopped(val reference: ActorReference<*>) : Protocol
-
-}
-
